Remove redundant @CorePlatformApi from DexFile.
This is unnecessary since the class is part of the public API (see
https://developer.android.com/reference/dalvik/system/DexFile) because
it has no @hide. The effect of this annotation is to cause the class's
public methods to appear in the core_platform_api current.txt, which
is misleading.

(There are members of this class which are marked @hide and
@CorePlatformApi. This is presumably intentional, and this change does
not affect them, they remain part of the core platform API.)

Test: treehugger
Bug: 154796679
Change-Id: Ie4c8ce97789c639f321c8cbb283b1f259183bf98